Revert "Refactor kubectl methods into utility class, so that we can easily integrate new tests and kube files written in json or other formats."
